requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
apscheduler==3.6.3
cachetools==4.2.2; python_version ~= '3.5'
certifi==2021.10.8
charset-normalizer==2.0.9; python_version >= '3'
click==8.0.3; python_version >= '3.6'
flask==2.0.2
futures==2.2.0
goslate==1.5.2
idna==3.3; python_version >= '3'
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.3; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
pytelegrambotapi==4.2.2
python-dotenv==0.19.2
python-telegram-bot==13.9
pytz-deprecation-shim==0.1.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pytz==2021.3
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
setuptools==60.0.3; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
tornado==6.1; python_version >= '3.5'
tzdata==2021.5; python_version >= '3.6'
tzlocal==4.1; python_version >= '3.6'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.2; python_version >= '3.6'
